+++ plugins/amazonq/codewhisperer/jetbrains-community/tst/software/aws/toolkits/jetbrains/services/codewhisperer/CodeWhispererCodeCoverageTrackerTest.kt
@@ -64,7 +64,6 @@ import software.aws.toolkits.jetbrains.services.codewhisperer.telemetry.CodeCove
 import software.aws.toolkits.jetbrains.services.codewhisperer.telemetry.CodeWhispererCodeCoverageTracker
 import software.aws.toolkits.jetbrains.services.codewhisperer.toolwindow.CodeWhispererCodeReferenceManager
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererConstants.TOTAL_SECONDS_IN_MINUTE
-import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererUtil.getUnmodifiedAcceptedCharsCount
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.CrossFileStrategy
 import software.aws.toolkits.jetbrains.services.telemetry.NoOpPublisher
 import software.aws.toolkits.jetbrains.services.telemetry.TelemetryService
@@ -439,50 +438,6 @@ internal class CodeWhispererCodeCoverageTrackerTestPython : CodeWhispererCodeCov
         sut.forceTrackerFlush()
 
         verify(sut, Times(0)).emitCodeWhispererCodeContribution()
-    }
-
-    @Test
-    fun `test getUnmodifiedAcceptedCharsCount()`() {
-        var originalRecommendation = "foo"
-        var modifiedRecommendation = "fou"
-        var unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo(2)
-
-        originalRecommendation = "foo"
-        modifiedRecommendation = "f11111oo"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo(3)
-
-        originalRecommendation = "foo"
-        modifiedRecommendation = "fo"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo(2)
-
-        originalRecommendation = "helloworld"
-        modifiedRecommendation = "HelloWorld"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo("helloworld".length - 2)
-
-        originalRecommendation = "helloworld"
-        modifiedRecommendation = "World"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo("helloworld".length - "hello".length - 1)
-
-        originalRecommendation = "CodeWhisperer"
-        modifiedRecommendation = "CODE"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo(1)
-
-        originalRecommendation = "CodeWhisperer"
-        modifiedRecommendation = "codewhispererISBEST"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo("CodeWhisperer".length - 2)
-
-        val pythonCommentAddedByUser = "\"\"\"we don't count this comment as generated by CodeWhisperer\"\"\"\n"
-        originalRecommendation = "x, y):\n\treturn x + y"
-        modifiedRecommendation = "x, y):\n$pythonCommentAddedByUser\treturn x + y"
-        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
-        assertThat(unmodifiedCharsCount).isEqualTo(originalRecommendation.length)
     }
 
     @Test
