+++ plugins/amazonq/codewhisperer/jetbrains-community/tst/software/aws/toolkits/jetbrains/services/codewhisperer/CodeWhispererUtilTest.kt
@@ -23,6 +23,7 @@ import software.aws.toolkits.jetbrains.core.credentials.sono.SONO_URL
 import software.aws.toolkits.jetbrains.core.region.MockRegionProviderRule
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererUtil.getCompletionType
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererUtil.getTelemetryOptOutPreference
+import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererUtil.getUnmodifiedAcceptedCharsCount
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.runIfIdcConnectionOrTelemetryEnabled
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.toCodeChunk
 import software.aws.toolkits.jetbrains.settings.AwsSettings
@@ -192,4 +193,49 @@ class CodeWhispererUtilTest {
         assertThat(AwsSettings.getInstance().isTelemetryEnabled).isFalse
         assertThat(getTelemetryOptOutPreference()).isEqualTo(OptOutPreference.OPTOUT)
     }
+
+    @Test
+    fun `test getUnmodifiedAcceptedCharsCount()`() {
+        var originalRecommendation = "foo"
+        var modifiedRecommendation = "fou"
+        var unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo(2)
+
+        originalRecommendation = "foo"
+        modifiedRecommendation = "f11111oo"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo(3)
+
+        originalRecommendation = "foo"
+        modifiedRecommendation = "fo"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo(2)
+
+        originalRecommendation = "helloworld"
+        modifiedRecommendation = "HelloWorld"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo("helloworld".length - 2)
+
+        originalRecommendation = "helloworld"
+        modifiedRecommendation = "World"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo("helloworld".length - "hello".length - 1)
+
+        originalRecommendation = "CodeWhisperer"
+        modifiedRecommendation = "CODE"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo(1)
+
+        originalRecommendation = "CodeWhisperer"
+        modifiedRecommendation = "codewhispererISBEST"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo("CodeWhisperer".length - 2)
+
+        val pythonCommentAddedByUser = "\"\"\"we don't count this comment as generated by CodeWhisperer\"\"\"\n"
+        originalRecommendation = "x, y):\n\treturn x + y"
+        modifiedRecommendation = "x, y):\n$pythonCommentAddedByUser\treturn x + y"
+        unmodifiedCharsCount = getUnmodifiedAcceptedCharsCount(originalRecommendation, modifiedRecommendation)
+        assertThat(unmodifiedCharsCount).isEqualTo(originalRecommendation.length)
+    }
+
 }
